Ignore surrounding spaces around the "CodeStyle" and "RuleExecution" values in the '.editorconfig' file
The default EnumValueParser of ec4j does not ignore such spaces which incorrectly leads to values not parsed correctly. Additional trailing space typically result from EOL comments in the '.editorconfig' file. For example:
`codestyle = official # some comment`.
The comment (including the hash sign) is ignored correctly but the space which separates the value and the hash sign is added to the value itself.