Add system property CLI_TEST_MAX_DURATION_IN_SECONDS
The Github CICD build is typically slower than the local machine. As of that it might need more time to complete a single CLI tests. For a developer it is usually not important whether the CICD build takes 5 versus 10 minutes. On the local machine it is annoying when the unit tests take long due because the developer is waiting for the tests to complete. So on the local machine it is better to fail the test after a shorter timeout.