Replace typealias EditorConfigProperties with data class EditorConfig
The typealias EditorConfigProperties was not used consistently in the KtLint code base. Some methods used the underlying data type `Map<String, Property>` instead. The `EditorConfig` enforces a consistent usage as it fully encapsulates the `Map<String, Property>` data type. Also, the `EditorConfig` has a secondary constructor with a vararg arguments which results in better readable code.

The interface UsesEditorConfigProperties which optionally could be implemented on a Rule class is replaced with a property in the Rule class. If a rule wants to read a property from the EditorConfig, the property has to be registered in the field `usesEditorConfigProperties` of the Rule class. The RuleExecution context which provides the EditorConfig in the call to the `beforeFirstNode` function of the Rule instance only provides the properties that are registered in the field `usesEditorConfigProperties`.

The "end_of_line" property has been transformed to an official EditorConfigProperty `END_OF_LINE_PROPERTY` and has been added to the default editor config properties. The value "native" has been dropped as this value is not recognized by the ec4j library nor IntelliJ IDEA.

The EditorConfigGenerator has been simplified as the EditConfig can not contain multiple properties with the same name but having different Property definitions.

Fixed a problem in `FORCE_MULTILINE_WHEN_PARAMETER_COUNT_GREATER_OR_EQUAL_THAN_PROPERTY` the old default value -1 is not accepted as it is not a positive integer.

Fixed a problem in the EditorConfigLoader when `ignoreEditorConfigOnFileSystem` is enabled. The editorConfigDefaults will only be taken into account whenever the ec4j library is queried to retrieve the editor config for an extension at an existing path. This flag is supposed to be enabled while running unit tests which should not be affected by a '.editorconfig' file which exists on the filesystem where the tests are executed. For now, it is assumed that the file system does not contain such file in the root of the file system.

Create extension function to transform an EditorConfigProperty to an ec4j Property with a value.