Avoid class cast exceptions when retrieving a property without (ec4j) type
Both the rules provided by Ktlint and custom rules can define properties with property types that are unknown to ec4j. If such custom property types are not provided to the ec4j library when reading the ".editorconfig" file this results in a property with a null type. Parsing a value of such a property always results in a value of type String which might not corresponding with the real type of the property.

Another problem is that ec4j expects the name of a property to be identical to the name of the corresponding PropertyType. When no propertyType is found with a name identical to the name of the property, the type of the property becomes null.

When building a property using the "Property.Builder" of ec4j, it is not possible to create a property having a null type. But when loading an ".editorconfig" file containing a custom property for which the propertyType was not registered in the EditorConfigLoader of ec4j does result in such a property.

Problems above are mitigated as follows:
* Wrap the ec4j EditorConfigLoader into EditorConfigLoaderEc4j and ensure that only the latter class is calling the ec4j EditorConfigLoader.
* Enforce that the EditorConfigDefaults can only be loaded when the propertyTypes have been given.
* When retrieving a value for an EditorConfigProperty from EditorConfig then always ignore the type of the ec4j property but use the type of the EditorConfigProperty itself.
* Add KtlintTestFileSystem as test utility whenever testing with ".editorconfig" files. Refactor all tests doing so.