Replace KtLintRuleEngine parameter "ignoreEditorConfigOnFileSystem" with "fileSystem"
This property is primarily intended to be used in unit tests. By specifying an alternative FileSystem the unit test gains control on whether the EditorConfigLoader should or should not read specific ".editorconfig" files. For example, it is considered unwanted that a unit test is influenced by the ".editorconfig" file of the project in which the unit test is included.