Do not expose rule runners from RuleExecutionContext in case the rule is disabled.
After loading the EditorConfig it can be determined which rules are enabled or disabled. The "EditorConfig.get(EditorConfigProperty)" function can not be used for this as this would require such properties to be created dynamically. As those properties are required only to decide which rules are executed, it does not make sense to store them in the EditorConfig that is used for propagation of properties to the rules.

The logic which determine whether a rule is enabled or disabled is moved from the VisitorProvider to the new class RuleRunnerFilter. The VisitorProvider now receives a set of RuleRunners that are enabled. The VisitorProvider class could further be simplified by removing logic regarding skipping of RuleRunners containing an AfterRunRule visitor modifier. This logic could not be triggered as the RuleRunnerSorter throws an exception in case no RuleRunners remains after eliminating rules that should not run.

TODO: Likely there is more logic that can be moved from the RuleRunnerSorter to other classes. The RuleRunnerSorter should only sort RuleRunners and not eliminating them.