Refactor RuleFilter
RuleFilters are a way to select the RuleRunners which are allowed to be passed to the next filter before finally being sorted and used for processing the files. Each RuleFilter class has a single responsibility. The RuleFilter classes are invoked via the RuleExecutionContext class so that the order of the filter remains clear.