Replacing a whitespace containing a newline after the return type may not result in a max-line-length violation
In case the whitespace contains a newline than replacing it with a single space results in merging the lines to a single line. This rule allows this only when the merged lines entirely fit on a single line.

Suppose that code below does not fit on a single line than this rule no longer emits a violation:
    fun foo():
        String = "some-looooooooooooooooong-string"

See FunctionSignatureRule for reformatting the function like below:
    fun foo(): String =
        "some-looooooooooooooooong-string"

Closes #1764