Move Baseline to ktlint-cli-reporter-baseline (#2013)
The baseline reporter creates the "baseline.xml" file which is read by the Baseline class. By moving the Baseline class into the reporter module, both the producer and consumer of the "baseline.xml" are located in same module which encapsulates this functionality. Also, it makes it possible for API consumers to reuse the functionality *and* to keep in sync with the Ktlint CLI.