Disable extension point org.jetbrains.kotlin.com.intellij.treeCopyHandler (#2044)
* Allow to disable registration of extension point `org.jetbrains.kotlin.com.intellij.treeCopyHandler` in the kotlin embeddable compiler to investigate the impact whenever this extension point will not be supported in Kotlin 1.9

* Replace `LeafElement.replaceWithText(String)` with `LeafElement.rawReplaceWithText(String)`

Up until Kotlin 1.8 the KotlinPsiFactory registered the treeCopyHandler extension point which is needed in order to use the `LeafElement.replaceWithText(String)`. With the Kotlin 1.9 it is no longer possible to register this extension point which results in exception.

The exception does not occur when using method `LeafElement.rawReplaceWithText(String)`.

* Replace `PsiElement.addAfter(PsiElement, PsiElement)` with `AstNode.addChild(AstNode, AstNode)`

Up until Kotlin 1.8 the KotlinPsiFactory registered the treeCopyHandler extension point which is needed in order to use the `PsiElement.addAfter(PsiLeaf, PsiLeaf)`. With the Kotlin 1.9 it is no longer possible to register this extension point which results in exception.

The exception does not occur when using method `AstNode.addChild(AstNode, AstNode)`. Note that this method inserts the new node (first) argument *before* the second argument node and as of that is not a simple replacement of the `PsiElement.addAfter(PsiElement, PsiElement)`.

* Refactor - Keep main element as type ASTNode and convert to psi when needed

* Refactor - Split test and add some test cases

* Replace `PsiElement.replace(PsiElement)` with a sequence of `AstNode.addChild(AstNode, AstNode)` and `AstNode.removeChild(AstNode)`

Up until Kotlin 1.8 the KotlinPsiFactory registered the treeCopyHandler extension point which is needed in order to use the `PsiElement.replace(PsiElement)`. With the Kotlin 1.9 it is no longer possible to register this extension point which results in exception.