Run format up to 3 times in case formatting introduces changes which also can be autocorrected
Closes #2084

Do not require a blank line between the class/method signature and a nested declaration in rule blank-line-before-declaration.

Allow branches with empty blocks in if-else.

Emit violation which can not be autocorrected in case wrapping the operator reference in a binary expression would still violate the max-line-length.

Only emit violation in if-else-wrapping rule in case a newline is missing. An incorrect indent should not lead to a violation about a missing newline.

Swap the order in which the binary-expression-wrapping and argument-list-wrapping rules run. Binary expressions containing call expressions, or vice versa, are not autocorrected as it needs manual intervention whether it is better to wrap the operation reference or the argument value.