Add rule to migrate away from ktlint-disable directives (#2068)
* Add rule to migrate ktlint-disable directives from comments to @Suppress or @SuppressWarnings annotations

Note that the SuppressionLocatorBuilder never ignore this rule as otherwise the suppression hint to ignore *all* rules would also ignore the rule which has to migrate this suppression hint.

Closes #1947

* Fix max line length for markdown files

* Fix wildcard imports settings

* Improve trace logging by printing a lint error directly when it is emitted

The reports print the lint violations in which they occur in the source code. This is not necessarily the order in which they are detected and fixed. For debugging some type of problems, the latter order is important.

* Add "libs.logback" as runtime only dependency so that logging is visible in unit tests