Fix AST when adding annotation on an expression
Whenever an expression is annotated, the expression needs to be wrapped inside an ANNOTATED_EXPRESSION element which also contains the ANNOTATION_ENTRY. Note, the code previously generated resulted in a correct AST if the code was reloaded. But without this fix, the SuppressionLocatorBuilder did not process the newly added annotation and as of that resulted in changing the AST by rules running after the KtlintSuppressionRule which is forced to be run first.