Rewrite a multiline function literal containing a single expression in the block to a single line when it will not exceed the max line length
The EditorConfig now also contains the set of RuleIds that are enabled for the file which is being processed. This allows rules to conditionally skip part of their functionality in case another rule is enabled.

Wrapping of the function literal in `wrapping` and `statement-wrapping` rules is disabled whenever the `function-literal` is enabled to avoid competing and contradictory changes.

Closes #2181