Ignore override of function in rule `function-naming`
A function override should not be reported as the interface of class that defines the function might be out of scope of the project in which case the function name can not be changed. Note that the function will still be reported at the interface or class itself whenever that interface or class is defined inside the scope of the project.

Closes #2271