WIP: add maven relocation block to `ktlint-cli` and `ktlint-cli-reporter-baseline`
When publishing to the local Maven repository, a `0.51.0-FINAL` version of the `pom.xml` is created for the modules above. The modules contain the expected content. But the `pom.xml` is missing the signature. Another side effect is that the `1.0.1-SNAPSHOT` version of the modules above is not created anymore as the `signing` block is now added to the module.

In 1.0.0 more modules have been relocated. They need to be added as well when it is clear how this should be approached.