Add maven relocation (#2282)
In the `1.0.0` release, the maven coordinates of some modules were changed. To support migration to the new coordinates a special relocatation `pom.xml`s (with version `0.51.0-FINAL`) are created. Builds running on a release up until `0.50.x` that will upgrade to the `0.51.0-FINAL` release will then get a message with the new maven coordinates.

Closes #2280