Remove dependency on `discouraged-comment-location` rule
The `discouraged-comment-location` rule contained several unrelated comment locations. When disabling the rule, ktlint might throw an exception when another rule depends on it. It is understandable that users do not instantly understand the relation between for example the `if-else-wrapping` and the `discouraged-comment-location` rule. So rules now have to emit warnings themselves whenever comments can not be processed.

For the `type-argument/type-argument-list`, `type-projection/type-parameter-list`, `value-argument/value-argument-list`, `value-parameter/value-parameter-list` new rules have been created to disallow comments. As those rules are more specific, it is more reasonable why specific rules depend on those rules.