Fix operator spacing
If an operation reference contains an identifier, then it is used in a DSL notation in which the identifier should be separated with a from a parenthesized expression to avoid confusion with a function call. E.g. format as `every { foo() } returns (bar) andThen (baz)`

Offsets for operators having length 2 or more, and not followed by single space were incorrect as length of the operator was not taken into account.

Closes #2459