Prevent exception in `binary-expression-wrapping` rule in case it is preceded on the same line by a block comment containing a newline character
Function `leavesOnLine` calls function `getFirstLeafOnLineOrSelf` which finds the first leaf on the line by looking for newline character. A block comment containing a newline character resulted in considering that block comment to be the first element on the line, even in case it was preceded by other elements on the same line.

Due to this change, the `max-line-length` rule failed, which has been resolved by refactoring the calculation of the line length.

Closes #2601