Add RuleAutocorrectApproveHandler interface
If a rule implements this interface, and changes the methods beforeVisitChildNodes and/or afterVisitChildNodes in compliance with this interface then the rule is able to support fixing of individual violations by requesting approval to autocorrect each violation found. For this the handler needs to be implemented by the API Consumer that is calling the format function of the KtLintRuleEngine.

For now this is implemented in a separate interface to prevent breaking changes in the Rule 1.x API.