Check that code can still be parsed after a format
In rare cases it can happen that formatted code can not be compiled anymore. Add additional checking in `KtLintAssertThat` that linting of formatted code still succeeds during unit tests. For release testing, the CLI has been expanded with a new (hidden) option that forces lint to run after format, and to stop execution as soon as the formatted code can not be compiled anymore.

Note that this still does not guarantee 100% that formatted code will be compilable. The additional tests will only ensure that they won't go unnoticed when the problem occurs in unit tests, or in the sample projects that are used for release testing.

Closes #2691