Refactor KotlinPsiFileFactory to KotlinCompiler (#2951)
* Refactor KotlinPsiFileFactory to KotlinCompiler

Remove workaround for Kotlin Compiler version 1.6 on Java 1.8 (see #1271) which requires KtlintRuleEngine to be injected. The problem of #1271, can still be reproduced with Ktlint 0.43.0 on java 1.8.0_432. With Ktlint 0.44.0 until 1.5.0 which included the workaround, the problem can not be reproduced on that same java version. Without the workaround on that same java version, the problem is also resolved. So, the workaround is no longer needed as it has been resolved in a Kotlin version after 1.6.0.

By removing this workaround, the code is simplified.

* Refactor KotlinPsiFileFactory to KtlintKotlinCompiler

Remove workaround for Kotlin Compiler version 1.6 on Java 1.8 (see #1271) which requires KtlintRuleEngine to be injected. The problem of #1271, can still be reproduced with Ktlint 0.43.0 on java 1.8.0_432. With Ktlint 0.44.0 until 1.5 which included the workaround, the problem can not be reproduced on that same java versionn. Without the workaround on that same java version, the problem is also resolved. So, the workaround is no longer needed as it has been resolved in a Kotlint version after 1.6.0.

By removing this workaround, the code is simplified.

* Refactor PsiFileFactory to KtlintKotlinCompiler

This reduces the number of conversions between "psi" and ASTNode. Also, the code is simplified