Set system properties to prevent failure when running in ktlint-intellij-plugin with custom installation (#2970)
Ktlint contains the embedded Kotlin compiler. When running in a process in Intellij IDEA, the initialization of the compiler requires some system properties to be set. When not set, this results in the exception reported, but only when running as (sub)process in context of Intellij IDEA. When using as standalone version, the exception is not run (because it is not running inside the IDEA).

Setting the properties to a non-null value during initialization of the embedded Kotlin compiler, prevents the exception from being thrown.

Closes #2969