Fix slow response of the git pre commit hook as generated by ktlint
When no kotlin files have been changed the list of patterns provided via stdin by the 'git diff' command is empty. Running ktlint then takes considerably longer as the entire project directory is traversed looking for kotlin files. But no such files will be found as the list of provided patterns is empty. So the project tree should only be traversed for a non-empty list of patterns.

Closes 2976