Fix `function-signature` rule for context parameters
Starting from Kotlin 2.2.0 the parser in the embedded compiler is able to process both context receiver (deprecated) and context parameter. Note that in a real project the compiler flag "-Xcontext-receiver", -or- "-Xcontext-parameter" needs to be set before the code successfully can be compiled. For the parsing only this is not required.

However, upgrading the kotlin version to 2.2.0 did break a unit test. A context receiver placed on a separate line before the function was no longer properly recognized.