Prevent test failures on Windows runners in GitHub build
Especially on Windows runners it sometimes happens that the test process is not terminated after completion of the command that is being tested. If the command prints its exit code to the log, that is being used once the test process times out. Commands which are fully handled by Clikt, do not print the exit code to the log, which results in a test failure if the test process expires. So, for Clikt commands the assertions on the exit code are ignored, if the test process has timed out.